Add files via upload
added a few fixes for weird issues. There seems to be a problem with pytube's streamingData key so I added a try except to be able to catch that. Also noticed that some videos which have problematic windows file characters often get download issues so I added a fix for the most common ones (namely these ones ->  "   :   ?   |   . ) 
I am planning on rewriting a decent portion of the script in the somewhat near future but it's probably not gonna happen until the 13th of may as I'm gonna take a short break from this (and others) project